Renamed files, improved conditions for warning
