+++ test/plots/index.js
@@ -1,4 +1,3 @@
-export {default as facet_warning} from "./facet_warning.js";
 export {default as aaplBollinger} from "./aapl-bollinger.js";
 export {default as aaplCandlestick} from "./aapl-candlestick.js";
 export {default as aaplChangeVolume} from "./aapl-change-volume.js";
@@ -63,6 +62,7 @@ export {default as empty} from "./empty.js";
 export {default as emptyLegend} from "./empty-legend.js";
 export {default as emptyX} from "./empty-x.js";
 export {default as energyProduction} from "./energy-production.js";
+export {default as facetWarning} from "./facet-warning.js";
 export {default as faithfulDensity} from "./faithful-density.js";
 export {default as faithfulDensity1d} from "./faithful-density-1d.js";
 export {default as figcaption} from "./figcaption.js";
